+++ tests/test_cli_publish.py
@@ -26,14 +26,17 @@ def test_cli_publish(mock_request_post):
     runner = CliRunner()
 
     # sends expected request
-    mock_request_post.return_value = MockResponse({"message": "mock message"}, 200)
+    mock_request_post.return_value = MockResponse(
+        {"message": "mock message", "jobId": "1234fakejob"}, 200
+    )
     result = runner.invoke(publish, ["test.id"])
     mock_request_post.assert_called_with(
         "https://api.mapbox.com/tilesets/v1/test.id/publish?access_token=pk.eyJ1IjoidGVzdC11c2VyIn0K"
     )
     assert result.exit_code == 0
+    print(result.output)
     assert (
-        "You can view the status of your tileset with the `tilesets status test.id` command."
+        '{"message": "mock message", "jobId": "1234fakejob"}\n\n✔ Tileset job received. Visit https://studio.mapbox.com/tilesets/test.id or run tilesets job test.id 1234fakejob to view the status of your tileset.\n'
         in result.output
     )
 
@@ -42,7 +45,9 @@ def test_cli_publish(mock_request_post):
 @mock.patch("requests.Session.post")
 def test_cli_publish_use_token_flag(mock_request_post):
     runner = CliRunner()
-    mock_request_post.return_value = MockResponse({"message": "mock message"}, 200)
+    mock_request_post.return_value = MockResponse(
+        {"message": "mock message", "jobId": "1234fakejob"}, 200
+    )
     # Provides the flag --token
     result = runner.invoke(publish, ["test.id", "--token", "flag-token"])
     mock_request_post.assert_called_with(
@@ -50,6 +55,6 @@ def test_cli_publish_use_token_flag(mock_request_post):
     )
     assert result.exit_code == 0
     assert (
-        "You can view the status of your tileset with the `tilesets status test.id` command."
+        '{"message": "mock message", "jobId": "1234fakejob"}\n\n✔ Tileset job received. Visit https://studio.mapbox.com/tilesets/test.id or run tilesets job test.id 1234fakejob to view the status of your tileset.\n'
         in result.output
     )